Split resolver behavior impl method tests
