+++ src/typechecker/tests/resolver_collection/behavior_impl_methods/mod.rs
@@ -1,5 +1,7 @@
 use super::*;
 
+mod restored_signatures;
+
 #[test]
 fn collect_declarations_with_symbols_uses_resolver_impl_behavior_for_defaults() {
     let mut program = parse_program(
@@ -408,252 +410,4 @@ Point.implements(Json) {
         !tc.methods.contains_key("Point.encode"),
         "behavior impl method signatures should be owned by the behavior impl metadata pass"
     );
-}
-
-#[test]
-fn collect_declarations_with_symbols_does_not_fallback_to_stale_ast_behavior_impl_method_signature()
-{
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.set_return_type_for_test(Namespace::Value, "Point.encode", None);
-    if let Declaration::ImplBlock { methods, .. } = &mut program.declarations[2] {
-        if let Declaration::Function {
-            params,
-            return_type,
-            ..
-        } = &mut methods[0]
-        {
-            params[0].ty = AstType::Named("Stale".to_string());
-            *return_type = Some(AstType::Named("AlsoStale".to_string()));
-        }
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    assert!(
-            !tc.methods.contains_key("Point.encode"),
-            "resolver-backed behavior impl collection should not keep AST-only method metadata when resolver signature metadata is incomplete"
-        );
-}
-
-#[test]
-fn collect_declarations_with_symbols_clears_stale_behavior_impl_method_signature_after_key_restore()
-{
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.set_return_type_for_test(Namespace::Value, "Point.encode", None);
-    if let Declaration::ImplBlock {
-        type_name, methods, ..
-    } = &mut program.declarations[2]
-    {
-        *type_name = "Missing".to_string();
-        if let Declaration::Function {
-            name,
-            params,
-            return_type,
-            ..
-        } = &mut methods[0]
-        {
-            *name = "missing".to_string();
-            params[0].ty = AstType::Named("Stale".to_string());
-            *return_type = Some(AstType::Named("AlsoStale".to_string()));
-        }
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    assert!(
-        !tc.methods.contains_key("Missing.missing"),
-        "resolver-backed behavior impl collection should not keep stale AST method keys"
-    );
-    assert!(
-            !tc.methods.contains_key("Point.encode"),
-            "resolver-backed behavior impl collection should clear restored method keys when resolver signature metadata is incomplete"
-        );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_impl_method_signature_target_and_name_metadata(
-) {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::ImplBlock {
-        type_name, methods, ..
-    } = &mut program.declarations[2]
-    {
-        *type_name = "Missing".to_string();
-        if let Declaration::Function {
-            name,
-            params,
-            return_type,
-            ..
-        } = &mut methods[0]
-        {
-            *name = "missing".to_string();
-            params[0].ty = AstType::Named("Stale".to_string());
-            *return_type = Some(AstType::Named("AlsoStale".to_string()));
-        }
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let info = tc.methods.get("Point.encode").expect("impl method info");
-    assert!(!tc.methods.contains_key("Missing.missing"));
-    assert_eq!(info.params[0].0, "value");
-    assert_eq!(info.params[0].1, AstType::Named("Point".to_string()));
-    assert_eq!(info.return_type, AstType::Str);
-    assert!(
-        tc.diagnostics.is_empty(),
-        "resolver-restored behavior impl method signature should avoid stale AST diagnostics: {:?}",
-        tc.diagnostics
-    );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_impl_generic_method_template_target_and_name_metadata(
-) {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode<T> = (value: Point) str { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::ImplBlock {
-        type_name, methods, ..
-    } = &mut program.declarations[2]
-    {
-        *type_name = "Missing".to_string();
-        if let Declaration::Function {
-            name,
-            params,
-            return_type,
-            ..
-        } = &mut methods[0]
-        {
-            *name = "missing".to_string();
-            params.pop();
-            *return_type = None;
-        }
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let template = tc
-        .generic_methods
-        .get("Point.encode")
-        .expect("generic behavior impl method template");
-    assert!(!tc.generic_methods.contains_key("Missing.missing"));
-    assert!(!tc.generic_methods.contains_key("Point.missing"));
-    assert_eq!(template.type_params, vec!["T".to_string()]);
-    assert_eq!(template.params.len(), 1);
-    assert_eq!(template.params[0].name, "value");
-    assert_eq!(template.params[0].ty, AstType::Named("Point".to_string()));
-    assert_eq!(template.return_type, Some(AstType::Str));
-    assert!(
-        tc.diagnostics.is_empty(),
-        "resolver-restored behavior impl generic template should avoid stale AST diagnostics: {:?}",
-        tc.diagnostics
-    );
-}
-
-#[test]
-fn collect_declarations_with_symbols_clears_stale_behavior_impl_generic_method_template_after_key_restore(
-) {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode<T> = (value: Point) str { "point" }
-}
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.set_return_type_for_test(Namespace::Value, "Point.encode", None);
-    if let Declaration::ImplBlock {
-        type_name, methods, ..
-    } = &mut program.declarations[2]
-    {
-        *type_name = "Missing".to_string();
-        if let Declaration::Function {
-            name,
-            params,
-            return_type,
-            ..
-        } = &mut methods[0]
-        {
-            *name = "missing".to_string();
-            params[0].ty = AstType::Named("Stale".to_string());
-            *return_type = Some(AstType::Named("AlsoStale".to_string()));
-        }
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    assert!(
-        !tc.generic_methods.contains_key("Missing.missing"),
-        "resolver-backed behavior impl collection should clear stale AST generic method templates"
-    );
-    assert!(
-            !tc.generic_methods.contains_key("Point.encode"),
-            "resolver-backed behavior impl collection should clear restored generic method templates when resolver signature metadata is incomplete"
-        );
 }
